// frontend/api/escrow/service.did.d.ts
export interface Item {
  'id' : bigint,
  'status' : ItemStatus,
  'listime' : bigint,
  'owner' : Principal,
  'name' : string,
  'description' : string,
  'currency' : { 'ICP' : null } |
    { 'ICET' : null },
  'image' : string,
  'itype' : Itype,
  'price' : bigint,
}
export type ItemStatus = { 'pending' : null } |
  { 'list' : null } |
  { 'sold' : null } |
  { 'unlist' : null };
export type Itype = { 'nft' : null } |
  { 'service' : null } |
  { 'merchandise' : null } |
  { 'other' : null } |
  { 'coin' : null };